Add Temp and Serial sensors, make the tests work
The tests now work with a mocked Context that implements the Reader
trait from the modbus library.

The TemperatureSensor definition is now uncommented and works with the
`Box<dyn Reader>` args as well, which allows it to work both in the
test environment as well as in the actual code.

The same thing is also the case for the SerialSensor struct, although
that doesn't have live definitions or tests yet..