Add a sensor probe workspace
One of the challenges with this project is knowing what the various
sensor values are supposed to represent, and debugging by looking at
what the sensors themselves are outputting.

This adds a simple workspace which allows me to make single requests
over the command line to see what various sensors have set in their
registers. That then allows to me to determine what's wrong with code
and fix it.