Remove redundant workspaces, add metric for faults
This change mostly is about removing workspaces, which turned out to
make things more complicated unnecessarily, and adding a metric for
faults, so that they are published to prometheus. To make them work
it's slightly different to the others in that each fault is its own
label. This is usually not something you want to do with Prometheus
because high cardinality of metric labels is bad for performance, but
a) it's a small number of metrics and b) it's worth it for faults,
which should hopefully not happen very often anyway.