Add new HTTP interface and integration test framework
In order to have sensors that can be written to, we also need an
interface for things to write to them. As a starting point, an HTTP
interface makes the most sense, since things that send HTTP requests
to change sensor values can be easily written and triggered with
cron or whatever.

In the future this will be expanded upon.