Improve code layout and versioning
This change moves around the code layout. Now the executable is built using the
code in 'bin', with the rest living in 'lib'. This should make code reuse easier.

I've also moved the integration tests into their own crate, which means they now
don't run twice and generally things are a lot neater.

I've also changed the versioning to <Year>.<Month>.<Version>. This works better
with the style of development and the type of project this is, since the goal is
to produce an executable that works, not for the code to be shared by crate (at least
not right now). It can be changed in the future if things change.